requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
openai
faiss-cpu
streamlit
transformers
torch
sentence-transformers
flake8
black
isort
pytest
pyyaml
playwright
pandas
tiktoken
python-dotenv
langchain
langchain-community
langchain-huggingface
huggingface_hub
langchain-huggingface
langchain_openai